refactor: 사용자와 매장 연관관계 설정
