+++ src/main/java/com/burntoburn/easyshift/dto/store/req/StoreCreateRequest.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.dto;
+package com.burntoburn.easyshift.dto.store.req;
 
 
 import lombok.AllArgsConstructor;
