+++ src/test/java/com/burntoburn/easyshift/store/StoreCreateCheckTest.java
@@ -1,6 +1,6 @@
 package com.burntoburn.easyshift.store;
 
-import com.burntoburn.easyshift.dto.StoreCreateRequest;
+import com.burntoburn.easyshift.dto.store.req.StoreCreateRequest;
 import com.burntoburn.easyshift.entity.store.Store;
 import com.burntoburn.easyshift.repository.store.StoreRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
